requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.1.10
django-cors-headers==3.0.2
django-countries==5.3.2
django-extensions==2.1.4
django-rest-knox==4.1.0
djangorestframework==3.9.1
ez-setup==0.9
factory-boy==2.11.1
Faker==1.0.2
Pillow==5.4.1
pycparser==2.19
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0
text-unidecode==1.2
